Add MarkType

+++ GeneralStructures.cs
@@ -402,7 +402,7 @@ namespace PatcherYRpp
 
         public int num_steps()
         {
-            return Math.Abs(this.difference()) / this.turn_rate();
+            return Math.Abs((int)this.difference()) / (int)this.turn_rate();
         }
 
         public DirStruct Value; // target facing
